refactor(query): isolate transaction state and RPC workflow
