+++ ydb/src/client_query/internal.rs
@@ -1,4 +1,5 @@
-use super::exec::{ClientExecContext, TransactionExecContext};
+use super::exec::ClientExecContext;
+use super::transaction::TransactionExecContext;
 
 pub(crate) enum ExecCoreRef<'a> {
     Client(&'a mut ClientExecContext),
