+++ ydb/src/grpc_wrapper/raw_query_service/transaction_control.rs
@@ -15,6 +15,10 @@ impl TransactionId {
     pub(crate) fn as_str(&self) -> &str {
         &self.0
     }
+
+    pub(crate) fn into_string(self) -> String {
+        self.0
+    }
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
